// src/types.ts
export interface Branch {
  id: string;
  name: string;
  address?: string;
  phone?: string;
  cnpj?: string;
  stateRegistration?: string;
  email?: string;
  responsible?: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: 'ADMIN' | 'THERAPIST';
}